fix: use Color struct literals in terminal tests
Color became a packed struct(u32) with tag+payload but tests still
passed bare hex integers (0xFF0000). Replace with Color.RED etc.

Also fix terminal_frame_ptr() return type: u32 → usize for 64-bit.

All 41 terminal_wasm tests pass.

+++ src/terminal_wasm.zig
@@ -109,7 +109,7 @@ export fn terminal_commit() u32 {
 }
 
 /// Get pointer to the last committed frame (in WASM linear memory)
-export fn terminal_frame_ptr() u32 {
+export fn terminal_frame_ptr() usize {
     return @intFromPtr(&frame_buf);
 }
